Add custom deserialization process for rated property of account state response.

+++ shared/src/commonMain/kotlin/com/example/moveeapp_compose_kmm/domain/account/favorite/TvAccountState.kt
@@ -1,3 +1,3 @@
 package com.example.moveeapp_compose_kmm.domain.account.favorite
 
-data class TvAccountState(val isFavorite: Boolean, val rating: Double?)
+data class TvAccountState(val isFavorite: Boolean, val rating: Float?)
